Updated `set_virtual_position' so that it actually updates the range of the scroll bar before explicitly setting its value. If this is not performed, it is possible that the value exceeded the current range as the resizing of the scroll bar was normally performed on idle. To reproduce:
add a large number of items to the grid
set the virtual position to the maximum virtual position

A further optimization is to make sure that the scroll bars are only updated
if their range really does not permit the value to be set. This however,
is for later versions.

Also updated `perform_vertical_computation' and
`perform_horizontal_computation' so that the idle actions are extended in the
case where there are no rows or columns. Without this, it is possible that
wiping out the grid may not correctly resize the area that was occupied by
the scroll bars.

SVN: 53584@branches/Eiffel_56