Fixed two bugs:
1. Clicking on a position within the displayed list and then pressing the
up key caused a crash. This was reported by Peter Gummer as bug 5235.

2. Pressing page up or page down when all rows were visible did not
scroll to the very top or end, but adjusted this by the common row scroll
count. Now it scrolls to the very top or bottom.

Paul has verified and tested this code as well.

SVN: 53654@branches/Eiffel_56