Reworking the excluded class list slightly.
The additions here are just trying to be a little more logical
in the groupings of the classes. For example, the ARRAY related
classes appear in a single spot.

We must include many of ARRAY's ancestors as excluded because
otherwise EiffelStudio complains that ARRAY isn't as it 
expected.

SVN: 81827@branches/eth/eve