Adding support for create procedures defined in a parent.
Creation clauses who cite features which are declared in a parent
were previously not handled correctly (no code generated in the child class).
This should now be fixed, albeit there is a small matter of generating the
separate type name which currently works, but I'm not sure how robustly.

SVN: 82052@branches/eth/eve