Scoop fixes: degree integration, usage of the processor tag attribute, and generic implicit type.
The compilation degree format had some preconditions to only allow certain
degree numbers (which we extend to include some degree 7, for SCOOP).

Also, since the processor tag doesn't exist, we guard it more carefully 
(it should never be used in situations where we know the type isn't
separate).

Lastly, since the implied formal generic type is now detachable separate ANY,
we have to account for this in the assign_finder, which was generating
calls to .implementation_ on basic types.

SVN: 84545@branches/eth/eve