Fixed various issues with the checking of validity of types in the FILTER_LIST. It was discovered when using the testing library which adds and then remove a bunch of classes without cleaning their content. Another issue was found with the refactoring tool when renaming a class used in a constraint. Unfortunately none of the problemes were easily reproducible but still I've created eweasel test#incr401, test#term195 and test#testing001 to show what we were trying to fix.
The actual fix is to perform the cleanup after all classes have been removed in `process_removed_classes' because if one of the supplier of the removed class is using the remove class there is no more associated class but there are parts of our code that will not check that properly and will crash.
For the refactoring problem, something similar was happening but we could only protect the call `check_constraints' to prevent the crash and mark the entry invalid.
Finally the definition of `internal_is_valid_for_class' for QAT as it did not follow the same pattern as `good_generics' and was crashing the compiler when a QAT was used in a system where the classes has been removed.

SVN: 85049@trunk