Adapted C code to match new FreeELKS implementation of file handling: - EIF_FILENAME is now the type of paths at the C level: wchar_t * on Windows and char * on UNIX. - All external routines have been prefixed with `eif_' to avoid conflicts at the C level.
This implied that most of the _16 implementation that have been added have been obsoleted by the new implementation.

Added as many routine XML documentation to make things easier to understand and to read.

Fixed a bug in store/retrieve with respect to transient attributes when using basic or general store.

In addition got rid of obsolete VMS implementation.

On the Eiffel side:
* simplified the FILE_32 classes and DIRECTORY_32 classes to inherit from the non_32 counterpart as they now handle unicode paths.

SVN: 89717@trunk