Fixed a design flaw.
The (wrong) idea was that updates of a collection can be modelled as a delete followed by an insert,
mostly because an incremental update is not efficient because ABEL doesn't know which items get deleted.
This kind of behaviour was implemented at a very high level - in fact the backends don't even know 
how to 'update' a collection.

However if a collection (e.g. a SPECIAL object) is shared by two objects problems will arise...
Due to the delete-insert cycle the database entry will get a new primary key. This means the 
reference/foreign key of (at least one of) the owners will get invalid.

This commit fixes the behaviour at the higher levels of the stack, and adds a regression test.
It does not, however, fix the bug for existing backends.

SVN: 93050@trunk