Several changes to improve performance:
*) For read-only queries, ABEL does no longer identify an object,
as the object cannot be used for a subsequent write operation
anyway.

*) Instead of using reflection, the attribute name and static type
gets cached in TYPE_METADATA.

*) ABSTRACT_MANAGER maintains a hash table to quickly map a handler
to a specific type.

*) BACKEND_OBJECT no longer stores PS_PAIR (an obsolete concept anyway)
and has a new function to directly access its hash table. This avoids
a "double-lookup" which first checks if the attribute is present and 
then retrieves the attribute.

*) QUERY_CURSOR now generates and passes along ARRAYED_LIST instead of
LINKED_LIST.

*) Some changes in DEFAULT_OBJECT_HANDLER and READ_MANAGER to avoid
a string comparison against "NONE".

SVN: 93739@trunk