Improved change made at rev#74209 due to the following issue: - Open EiffelStudio non-maximized on screen 1 and quit - Open EiffelStudio (it shows non-maximized on screen 1), maximize it on screen 2, and quit - Open EiffelStudio, it is shown on screen 1, not screen 2 as expected.
This was caused because we never used the maximized_x_position and maximized_y_position, to position the window before maximizing it. We also do not need the maximized width and height, so we removed them.

Refactored the preferences data of the window so that preferences and session data can share some code.

SVN: 93759@trunk