Eweasel tests for SCOOP exception handling.
The tests are written against the new semantics
by Benjamin Morandi, and currently they all fail.
Initially the tests were part of a naster thesis 
at ETH by Florian Besser.

The patch also includes a test for a bug
(test#scoop041) where the behaviour in molten
and frozen code differs when evaluating a
precondition on a controlled object.

SVN: 96579@trunk