Added test#scoop049 for a bug related to impersonation mechanism.
The bug occurs when two consecutive queries happen on the same
object. In that case, the second call does not perform lock passing.
The eweasel test case then fails with a deadlock.

It is caused by a buggy impersonation mechanism which only occurs
after the first synchronous call.

SVN: 96582@trunk