Added a C replacement for std::vector<T>.
The new C class is basically a set of preprocessor macros that define a 
the struct and member functions for a dynamically growing arraay for any
user-defined type. It is thus a type-safe alternative to data structures
handling a general (void*) type and uses less space for 
less-than-pointer-sized elements.

SVN: 96798@trunk