Changed the way a terminating SCOOP thread is marked.
Instead of setting the is_processor flag to false, the
logical_id is now set to NULL_PROCESSOR_ID. That way
the garbage collector can distinguish between a regular
thread and a SCOOP processor which is about to be destroyed.

This commit fixes test#scoop028, test#scoop029 and test#bench003.

SVN: 97202@trunk