* Changed the code generation to be more compact (about a 2 to 6% size reduction), and it is about the same speed. * Changed the way we save/restore stacks in the generated code when you have a routine with rescue. It is much simpler now and clearer too with less differences between the normal and rescue case. Basically in the rescue case we add calls to other macros, instead of changing the macros we are using. The biggest change is that now we do:
RTLI(x);
RTLR(0, &Current);
...
RTLIU(x);

the last call is actually updating the top of the `loc_set' stack. And the final macro RTLE simply update the top pointer of `loc_set', no if is used to find out if we need to go back one chunk or not. The big assumption is that `loc_set' is always allocated before being used which is a given (see local.c).

* To support the above and reduce code duplication for stack handling we are using some kind of template mechanism to generate from a single file multiple generic derivation of the same stack code. 
* Replaced all stack manipulations using those generic routines. There are still some custom made stack manipulation but they are far less than before and much easier to understand
* The new stack have a clear semantic of what to expect, and access to the top element is always done without a check if the stack was created and not empty.
* Added quite a few assertions too.

See review#5762110734729216

SVN: 97570@trunk