First step towards new compiler interface:
* Introduced local variables in RTS_SD for
region and processor ID.
* Changed all RTS_R* macros to take the
processor ID instead of RTS_PID (Current)
* Changed eif_log_call to take processor
and region ID of the client, instead of
the supplier PID (since that one is in
the call_data struct anyway).
* Adapted all RTS_C* macros according to
the change in eif_log_call and made sure
that they use the new local variables.

SVN: 97931@trunk