Switched to the new impersonation mechanism.
The new mechanism doesn't use a setjmp() buffer
any more for an impersonated separate call. This
should make things faster, but it also means
that exceptions will jump across processor
boundaries and one has to recover from them later.

Furthermore, the whole SCOOP runtime now
distinguishes between the processor and region
ID of a client, although in most cases it's
the processor ID that is relevant. This
distinction was necessary for the new
impersonation mechanism as well as the 
(future) implementation of real passive regions.

The commit currently has two known bugs:
* The once values are not swapped.
* The wrong region ID may be used when a separate
callback happens to a non-impersonable region.
I'll fix these issues in later commits. On the
other hand, eweasel test#scoop065 is now working.

Even though the patch is relatively small, the
changes in the SCOOP runtime are quite heavy,
which means that there could be unintentional
regressions.

SVN: 97949@trunk