Removed the field rt_private_queue::call_stack_msg.
Previously, the field was necessary because
during the time an eif_scoop_call_data structure
was popped off the message channel and the time
it was actually executed, the garbage collector
was allowed to run (due to some saving and
restoring of last_exception). This is no longer
the case, and therefore the code can be simplified.

SVN: 98192@trunk