Fixed a compiler/run-time bug related to RTAR that did not take into account that the code is generated like 	RTAR (Current, loc1); 	*(Current + ...) = RTCCL (loc1); to take care about possible cloning of expanded objects. One solution would be to change code generation to produce something like 	tr1 = RTCCL (loc1); 	RTAR (Current, tr1); 	*(Current + ...) = tr1; However this could make generated code larger and slower at run-time because tr1 is volatile and cannot be stored in a register.
The current solution is to check whether the source object is either new or expanded. In the latter case, it would mean object cloning that would produce a new object anyway.

This fixes test#runtime021, and, possibly, the issues reported by users when an expanded object is attached to a reference attribute.

SVN: 101201@trunk