Committing files that were not included in rev#103906 by mistake.
Supported multi-branch expressions that are similar to multi-branch instructions, but Then_part and Else_part are expressions. The type of a multi-branch expression is computed similarly to the type of conditional expressions and actual genertics of manifest arrays.
Unified AST and byte nodes of multi-branch instructions and expressions by factoring out common parent code and adding formal generic types in the common parents.
Removed a (non-standard) restriction of BYTE_LIST and EIFFEL_LIST to deal only with a reference actual generic parameter.
Cosmetics:
	- replaced regular loops with across loops;
	- replaced voidness tests with object tests;
	- replaced reverse assignments with object tests.

SVN: 103907@trunk