feat: Added the cloud user profile functionality.
For now basic, just include a wikitext "about" field, so that user can describe themselves, add projects information, and so on.
Added SHOP_CUSTOMER, to get internal information of a customer, mostly from stripe if this is used as payment system.
Updated various cloud views, to have link on that new user profile page.

SVN: 105109@trunk