Use the compiler (usually gcc) instead of specific ' ld' tool, and harmonize various unices config files.
It seems like it may be more portable to let the compiler call the linker instead of invoking it directly as the compiler will be able to figure out the right way to link to glibc, which can vary based on different Linux systems.

This suggestion was provided via https://support.eiffel.com/report_detail/19873 as a workaround (and cleaner solution) for "brew".

SVN: 107381@trunk