Write the GUID of PE CodeView Directoy Entry using a determinist algorithm computed with a hash of the file content.
Updated the CLI_FACTORY to create a reproducible_debug_directory entry
and implemented for COM and IL emitter.

Updated the DBG_WRITER interface with open_read and close_read of features
to update the pdb file content.

Updated the IL_EMITTER_DEBUG_WRITER to compute the GUID using a deterministic algorithm
using the first 16 bytes of the hash (SHA256) of the file content.
Compute the PDB_ID of the pdb_stream as the first 16 bytes of the Checksum +
the timestamp of the file (using 20 bytes).

SVN: 107628@trunk